moved fetchPythonException to PythonUtils
